[ds-graph] break down djikstra heap in files
